+++ backend/attendance/src/main/java/org/egov/repository/AttendeeRepository.java
@@ -1,7 +1,9 @@
 package org.egov.repository;
 
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.repository.querybuilder.AttendeeQueryBuilder;
 import org.egov.repository.rowmapper.AttendeeRowMapper;
+import org.egov.tracer.model.CustomException;
 import org.egov.web.models.AttendeeSearchCriteria;
 import org.egov.web.models.IndividualEntry;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +12,9 @@ import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import static org.egov.Constants.INVALID_TENANT_ID;
+import static org.egov.Constants.INVALID_TENANT_ID_MSG;
 
 @Repository
 public class AttendeeRepository {
@@ -26,9 +31,22 @@ public class AttendeeRepository {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public List<IndividualEntry> getAttendees(AttendeeSearchCriteria searchCriteria) {
+    /**
+     * This method fetches the list of attendees based on the search criteria provided.
+     *
+     * @param tenantId        The tenant ID for which the search is being performed.
+     * @param searchCriteria   The criteria used to filter the attendees.
+     * @return List<IndividualEntry> A list of IndividualEntry objects representing the attendees.
+     */
+    public List<IndividualEntry> getAttendees(String tenantId, AttendeeSearchCriteria searchCriteria) {
         List<Object> preparedStmtList = new ArrayList<>();
-        String query = queryBuilder.getAttendanceAttendeeSearchQuery(searchCriteria, preparedStmtList);
+        String query = null;
+        // Wrapped query construction in try-catch to handle tenant ID validation failure
+        try {
+            query = queryBuilder.getAttendanceAttendeeSearchQuery(tenantId, searchCriteria, preparedStmtList);
+        } catch (InvalidTenantIdException e) {
+            throw new CustomException(INVALID_TENANT_ID, e.getMessage());
+        }
         List<IndividualEntry> attendanceStaffList = jdbcTemplate.query(query, attendeeRowMapper, preparedStmtList.toArray());
         return attendanceStaffList;
     }