+++ backend/attendance/src/main/java/org/egov/repository/querybuilder/StaffQueryBuilder.java
@@ -1,15 +1,23 @@
 package org.egov.repository.querybuilder;
 
+import lombok.RequiredArgsConstructor;
+import org.egov.common.exception.InvalidTenantIdException;
+import org.egov.common.utils.MultiStateInstanceUtil;
 import org.egov.web.models.StaffSearchCriteria;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import org.springframework.util.ObjectUtils;
 
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
+import static org.egov.common.utils.MultiStateInstanceUtil.SCHEMA_REPLACE_STRING;
+
+@RequiredArgsConstructor
 @Component
 public class StaffQueryBuilder {
+
+    private final MultiStateInstanceUtil multiStateInstanceUtil;
 
     private static final String ATTENDANCE_STAFF_SELECT_QUERY = " SELECT stf.id, " +
             "stf.individual_id, " +
@@ -23,19 +31,20 @@ public class StaffQueryBuilder {
             "stf.createdtime, " +
             "stf.lastmodifiedtime, " +
             "stf.tenantid, " + "stf.stafftype " +
-            "FROM eg_wms_attendance_staff stf ";
+            "FROM %s.eg_wms_attendance_staff stf ";
 
-    public String getActiveAttendanceStaffSearchQuery(StaffSearchCriteria criteria, List<Object> preparedStmtList) {
-        StringBuilder query = new StringBuilder(getAttendanceStaffSearchQuery(criteria, preparedStmtList));
+    public String getActiveAttendanceStaffSearchQuery(StaffSearchCriteria criteria, List<Object> preparedStmtList) throws InvalidTenantIdException {
+        StringBuilder query = new StringBuilder( getAttendanceStaffSearchQuery(criteria, preparedStmtList));
         addClauseIfRequired(query, preparedStmtList);
         query.append(" stf.deenrollment_date is null ");
 
         return query.toString();
     }
 
-    public String getAttendanceStaffSearchQuery(StaffSearchCriteria criteria, List<Object> preparedStmtList) {
+    public String getAttendanceStaffSearchQuery(StaffSearchCriteria criteria, List<Object> preparedStmtList) throws InvalidTenantIdException {
         StringBuilder query = new StringBuilder(ATTENDANCE_STAFF_SELECT_QUERY);
-
+        String tenantId = criteria.getTenantId();
+        query = new StringBuilder(String.format(String.valueOf(query), SCHEMA_REPLACE_STRING));
         List<String> staffUserIds = criteria.getIndividualIds();
         if (staffUserIds != null && !staffUserIds.isEmpty()) {
             addClauseIfRequired(query, preparedStmtList);
@@ -50,7 +59,6 @@ public class StaffQueryBuilder {
             preparedStmtList.addAll(criteria.getRegisterIds());
         }
 
-        String tenantId = criteria.getTenantId();
         if (tenantId != null && !tenantId.isEmpty()) {
             addClauseIfRequired(query, preparedStmtList);
             query.append(" stf.tenantid IN (").append(createQuery(Collections.singletonList(tenantId))).append(")");
@@ -63,7 +71,8 @@ public class StaffQueryBuilder {
             query.append(" stf.stafftype = ? ");
             preparedStmtList.add(staffType);
         }
-        return query.toString();
+
+        return multiStateInstanceUtil.replaceSchemaPlaceholder(String.valueOf(query), tenantId);
     }
     private void addClauseIfRequired(StringBuilder query, List<Object> preparedStmtList) {
         if (preparedStmtList.isEmpty()) {